requirements.txt
# User Interface
streamlit==1.42.0
requests>=2.32.5

# Web Framework
fastapi==0.129.0
uvicorn[standard]==0.41.0
python-multipart==0.0.22

# AI & RAG Orchestration
groq==0.18.0
langchain==1.2.10
langchain-community==0.4.1
langchain-openai==1.1.10
langchain-core==1.2.13

# Vector Database & Embeddings
faiss-cpu==1.13.2
pypdf==6.7.1
python-dotenv==1.2.1